refactor: TableGroup Order 의존성 제거
