+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -1,11 +1,6 @@
 package kitchenpos.application;
 
-import static kitchenpos.domain.OrderStatus.COOKING;
-
-import kitchenpos.dao.OrderDao;
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.application.validator.TableGroupValidator;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.dto.TableGroupResponse;
 import kitchenpos.dto.TableGroupSaveRequest;
@@ -13,24 +8,16 @@ import kitchenpos.repository.TableGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
 @Transactional(readOnly = true)
 public class TableGroupService {
 
-    private static final List<String> NONE_UNGROUP_ORDER_STATUS = List.of(COOKING.name(), OrderStatus.MEAL.name());
-
-    private final OrderDao orderDao;
-    private final OrderTableDao orderTableDao;
+    private final TableGroupValidator tableGroupValidator;
     private final TableGroupRepository tableGroupRepository;
 
-    public TableGroupService(final OrderDao orderDao,
-                             final OrderTableDao orderTableDao,
+    public TableGroupService(final TableGroupValidator tableGroupValidator,
                              final TableGroupRepository tableGroupRepository) {
-        this.orderDao = orderDao;
-        this.orderTableDao = orderTableDao;
+        this.tableGroupValidator = tableGroupValidator;
         this.tableGroupRepository = tableGroupRepository;
     }
 
@@ -43,15 +30,6 @@ public class TableGroupService {
     @Transactional
     public void ungroup(final Long tableGroupId) {
         TableGroup savedTableGroup = tableGroupRepository.getById(tableGroupId);
-        List<Long> orderTableIds = toOrderTableIds(tableGroupId);
-        savedTableGroup.ungroup(() ->
-                orderDao.existsByOrderTableIdInAndOrderStatusIn(orderTableIds, NONE_UNGROUP_ORDER_STATUS));
-    }
-
-    private List<Long> toOrderTableIds(final Long tableGroupId) {
-        return orderTableDao.findAllByTableGroupId(tableGroupId)
-                .stream()
-                .map(OrderTable::getId)
-                .collect(Collectors.toList());
+        savedTableGroup.ungroup(tableGroupValidator.validate(savedTableGroup));
     }
 }
